Refactor studio into hook and components
Extract studio state and logic into a new useStudioState hook and replace large inline UI in App.tsx with composed components. Added Header, Sidebar, WriteView, StructureView, SettingsView, PrefsView and AiAssistant components, and removed many local state variables and storage helper functions from App.tsx. Wiring now passes handlers/state from useStudioState (including backup/save functions) to the new components, drastically slimming App.tsx and improving separation of concerns.
